+++ crates/miden-proving-service-client/build.rs
@@ -34,11 +34,10 @@ fn main() -> miette::Result<()> {
 // HELPER FUNCTIONS
 // ================================================================================================
 
-/// Copies the tx_prover.proto file from the root proto directory to the proto directory of this
-/// crate.
+/// Copies the proto file from the root proto directory to the proto directory of this crate.
 fn copy_proto_files() -> miette::Result<()> {
-    let src_file = format!("{REPO_PROTO_DIR}/tx_prover.proto");
-    let dest_file = format!("{CRATE_PROTO_DIR}/tx_prover.proto");
+    let src_file = format!("{REPO_PROTO_DIR}/proving_service.proto");
+    let dest_file = format!("{CRATE_PROTO_DIR}/proving_service.proto");
 
     fs::remove_dir_all(CRATE_PROTO_DIR).into_diagnostic()?;
     fs::create_dir_all(CRATE_PROTO_DIR).into_diagnostic()?;
@@ -50,16 +49,18 @@ fn copy_proto_files() -> miette::Result<()> {
 fn compile_tonic_client_proto() -> miette::Result<()> {
     let crate_root =
         PathBuf::from(env::var("CARGO_MANIFEST_DIR").expect("CARGO_MANIFEST_DIR should be set"));
-    let dst_dir = crate_root.join("src").join("tx_prover").join("generated");
+    let dst_dir = crate_root.join("src").join("proving_service").join("generated");
 
-    // Remove `tx_prover.rs` if it exists.
-    fs::remove_file(dst_dir.join("tx_prover.rs")).into_diagnostic().ok();
+    // Remove `proving_service.rs` if it exists.
+    // We don't need to check the success of this operation because the file may not exist.
+    let _ = fs::remove_file(dst_dir.join("std").join("proving_service.rs"));
+    let _ = fs::remove_file(dst_dir.join("nostd").join("proving_service.rs"));
 
     let out_dir = env::var("OUT_DIR").into_diagnostic()?;
     let file_descriptor_path = PathBuf::from(out_dir).join("file_descriptor_set.bin");
 
     let proto_dir: PathBuf = CRATE_PROTO_DIR.into();
-    let protos = &[proto_dir.join("tx_prover.proto")];
+    let protos = &[proto_dir.join("proving_service.proto")];
     let includes = &[proto_dir];
 
     let file_descriptors = protox::compile(protos, includes)?;
@@ -71,9 +72,9 @@ fn compile_tonic_client_proto() -> miette::Result<()> {
     build_tonic_client(&file_descriptor_path, &std_path, protos, includes, false)?;
     build_tonic_client(&file_descriptor_path, &nostd_path, protos, includes, true)?;
 
-    // Replace `std` references with `core` and `alloc` in `tx_prover.rs`.
+    // Replace `std` references with `core` and `alloc` in `proving_service.rs`.
     // (Only for nostd version)
-    let nostd_file_path = nostd_path.join("tx_prover.rs");
+    let nostd_file_path = nostd_path.join("proving_service.rs");
     let file_content = fs::read_to_string(&nostd_file_path).into_diagnostic()?;
     let updated_content = file_content
         .replace("std::result", "core::result")