+++ CHANGELOG.md
@@ -16,6 +16,9 @@
 - Made NoteFile and AccountFile more consistent (#1133).
 - [BREAKING] Implement most block constraints in `ProposedBlock` (#1123, #1141).
 - [BREAKING] Added the `get_block_timestamp` procedure to the `miden` library (#1138).
+- Added serialization for `ProposedBatch`, `BatchId`, `BatchNoteTree` and `ProvenBatch` (#1140).
+- Added `prefix` to `Nullifier` (#1153).
+- [BREAKING] Implemented a `RemoteBatchProver`. `miden-proving-service` workers can prove batches (#1142).
 - [BREAKING] Implement `LocalBlockProver` and rename `Block` to `ProvenBlock` (#1152).
 
 ## 0.7.2 (2025-01-28) - `miden-objects` crate only